refactor: redux to hooks
