+++ src/utils/MoviesApiRequest.js
@@ -24,6 +24,7 @@ class MoviesRequests {
     return fetch(`https://api.themoviedb.org/3/movie/upcoming?api_key=${process.env.REACT_APP_API_KEY}&language=en-US&page=1`)
     .then((response) => response.json());
   }
+
 }
 
 export default MoviesRequests;
